# Add benchmark project, some small optimizations