--- a/F23.StringSimilarity.Benchmarks/F23.StringSimilarity.Benchmarks.csproj
+++ b/F23.StringSimilarity.Benchmarks/F23.StringSimilarity.Benchmarks.csproj
@@ -1,0 +1,18 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <OutputType>Exe</OutputType>
+        <TargetFramework>net8.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <PackageReference Include="BenchmarkDotNet" Version="0.13.12" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\src\F23.StringSimilarity\F23.StringSimilarity.csproj" />
+    </ItemGroup>
+
+</Project>