--- a/F23.StringSimilarity.Benchmarks/Program.cs
+++ b/F23.StringSimilarity.Benchmarks/Program.cs
@@ -1,0 +1,4 @@
+﻿using BenchmarkDotNet.Running;
+using F23.StringSimilarity.Benchmarks;
+
+BenchmarkRunner.Run<Benchmarks>();